+++ frontend/src/components/Settings.js
@@ -162,7 +162,9 @@ function MatterSettings({ matterbridgeInfo, showSnackbarMessage }) {
   const [mdnsInterface, setmdnsInterface] = useState('');  
   const [ipv4Address, setIpv4Address] = useState('');  
   const [ipv6Address, setIpv6Address] = useState('');  
-  const [matterPort, setMatterPort] = useState(5540);  
+  const [matterPort, setMatterPort] = useState();  
+  const [matterDiscriminator, setMatterDiscriminator] = useState();  
+  const [matterPasscode, setMatterPasscode] = useState();  
 
   useEffect(() => {
     if (matterbridgeInfo.bridgeMode === undefined) return;
@@ -171,7 +173,9 @@ function MatterSettings({ matterbridgeInfo, showSnackbarMessage }) {
     setmdnsInterface(matterbridgeInfo.mattermdnsinterface);
     setIpv4Address(matterbridgeInfo.matteripv4address);
     setIpv6Address(matterbridgeInfo.matteripv6address);
-    setMatterPort(matterbridgeInfo.matterPort ?? 5540);
+    setMatterPort(matterbridgeInfo.matterPort);
+    setMatterDiscriminator(matterbridgeInfo.matterDiscriminator);
+    setMatterPasscode(matterbridgeInfo.matterPasscode);
   }, [matterbridgeInfo]);
 
   // Define a function to handle change debug level
@@ -217,6 +221,22 @@ function MatterSettings({ matterbridgeInfo, showSnackbarMessage }) {
     console.log('handleChangeMatterPort called with value:', event.target.value);
     setMatterPort(event.target.value);
     sendCommandToMatterbridge('setmatterport', event.target.value);
+    showSnackbarMessage('Restart Matterbridge to apply changes', 5);
+  };
+  
+  // Define a function to handle change matterDiscriminator
+  const handleChangeMatterDiscriminator = (event) => {
+    console.log('handleChangeMatterDiscriminator called with value:', event.target.value);
+    setMatterDiscriminator(event.target.value);
+    sendCommandToMatterbridge('setmatterdiscriminator', event.target.value);
+    showSnackbarMessage('Restart Matterbridge to apply changes', 5);
+  };
+
+  // Define a function to handle change matterPasscode
+  const handleChangemMatterPasscode = (event) => {
+    console.log('handleChangemMatterPasscode called with value:', event.target.value);
+    setMatterPasscode(event.target.value);
+    sendCommandToMatterbridge('setmatterpasscode', event.target.value);
     showSnackbarMessage('Restart Matterbridge to apply changes', 5);
   };
 
@@ -267,8 +287,28 @@ function MatterSettings({ matterbridgeInfo, showSnackbarMessage }) {
             }}/>
         </div>
         <div style={{ display: 'flex', alignItems: 'center', gap: '15px' }}>
-          <FormLabel color='readonly' style={{padding: '0px', margin: '0px'}}>Matter port:</FormLabel>
+          <FormLabel color='readonly' style={{padding: '0px', margin: '0px'}}>Matter commissioning port:</FormLabel>
           <TextField value={matterPort} onChange={handleChangeMatterPort} size="small" variant="outlined"
+            style={{ height: '30px', flexGrow: 1 }} InputProps={{
+              style: {
+                height: '30px',
+                padding: '0',
+              },
+            }}/>
+        </div>
+        <div style={{ display: 'flex', alignItems: 'center', gap: '15px' }}>
+          <FormLabel color='readonly' style={{padding: '0px', margin: '0px'}}>Matter commissioning discriminator:</FormLabel>
+          <TextField value={matterDiscriminator} onChange={handleChangeMatterDiscriminator} size="small" variant="outlined"
+            style={{ height: '30px', flexGrow: 1 }} InputProps={{
+              style: {
+                height: '30px',
+                padding: '0',
+              },
+            }}/>
+        </div>
+        <div style={{ display: 'flex', alignItems: 'center', gap: '15px' }}>
+          <FormLabel color='readonly' style={{padding: '0px', margin: '0px'}}>Matter commissioning passcode:</FormLabel>
+          <TextField value={matterPasscode} onChange={handleChangemMatterPasscode} size="small" variant="outlined"
             style={{ height: '30px', flexGrow: 1 }} InputProps={{
               style: {
                 height: '30px',